// product-detector-client/src/api/product-detector/models/product-detect-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ProductDetectResponseDto
 */
export interface ProductDetectResponseDto {
    /**
     * 
     * @type {Array&lt;ProductDto&gt;}
     * @memberof ProductDetectResponseDto
     */
    products?: any;
}