// product-detector-client/src/api/product-detector/models/image-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ImageDto
 */
export interface ImageDto {
    /**
     * 
     * @type {string}
     * @memberof ImageDto
     */
    fileName?: any;
    /**
     * 
     * @type {string}
     * @memberof ImageDto
     */
    format?: any;
    /**
     * 
     * @type {string}
     * @memberof ImageDto
     */
    dataEncoding?: any;
    /**
     * 
     * @type {string}
     * @memberof ImageDto
     */
    data?: any;
}